requirements.txt
streamlit
openai
pymupdf
langchain
langchain-community
langchain-openai==0.0.5
faiss-cpu
pillow
python-dotenv
requests
streamlit-lottie





l




